refactored it some more to work on the logic within the main function, where the main funtion will handle the arguments and conditions. still trying to figure out goroutines, as it will run really quick (0m4.746s on windows, failed to mention before), but won't output stdout/stderr as well. will look into it
